requirements.txt
streamlit
google-generativeai
python-dotenv
python-multipart
langchain
PyPDF2
chromadb
faiss-cpu
langchain_google_genai
langchain-community
langchain-openai
langchain-ollama